pyproject.toml
[project]
name = "an-hack-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.192.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.0",
    "playwright>=1.58.0",
    "watchdog>=6.0.0",
]
